Reverted checksum module separation
Functions are still separated however, and are part of the same module.
