+++ src/format/checksum/MessageChecksumValidator.c
@@ -4,14 +4,50 @@
 
 #include "proto_nmea/format/checksum/MessageChecksumValidator.h"
 
-#ifdef UNIT_TEST
-
-int8_t validateMessageChecksum_testWrapper(const unsigned char *message, size_t messageSize)
+static int8_t validateChecksumFormat(const unsigned char *checksumString)
 {
-    return validateMessageChecksum(message, messageSize);
+    // Should start with a checksum delimiter
+    if (checksumString[0] != PROTOCOL_CHECKSUM_DELIMITER)
+    {
+        if (strchr(((const char *) checksumString), PROTOCOL_CHECKSUM_DELIMITER))
+            return -EMISSING_CHECKSUM_DATA;
+        else
+            return -EMISSING_CHECKSUM;
+    }
+
+    // ToDo: Iterate with for loop
+    if (!isHex(checksumString[1]) || !isHex(checksumString[2]))
+        return -EINVALID_CHECKSUM;
+
+    return EVALID;
 }
 
-#endif
+static int8_t validateChecksumData(const unsigned char *checksumString, unsigned int calculatedChecksum)
+{
+    char receivedChecksum[3] = {0};
+
+    memcpy(receivedChecksum, checksumString + 1, 2);
+
+    int receivedChecksumValue = stringToHex((const unsigned char *) receivedChecksum);
+
+    if (calculatedChecksum != receivedChecksumValue)
+        return -EWRONG_CHECKSUM;
+
+    return EVALID;
+}
+
+static int8_t validateChecksumString(const unsigned char *buffer, unsigned int calculatedChecksum)
+{
+    uint8_t errorCode = validateChecksumFormat(buffer);
+    if (errorCode)
+        return errorCode;
+
+    errorCode = validateChecksumData(buffer, calculatedChecksum);
+    if (errorCode)
+        return errorCode;
+
+    return EVALID;
+}
 
 #ifdef UNIT_TEST
 
@@ -26,5 +62,15 @@ int8_t validateMessageChecksum(const unsigned char *message, size_t messageSize)
     int calculatedChecksum = calculateChecksum(message + MESSAGE_TYPE_START_INDEX,
                                                message + checksumDelimiterPosition);
 
-    return validateChecksum(message + checksumDelimiterPosition, calculatedChecksum);
+    return validateChecksumString(message + checksumDelimiterPosition, calculatedChecksum);
 }
+
+#ifdef UNIT_TEST
+
+int8_t validateMessageChecksum_testWrapper(const unsigned char *message, size_t messageSize)
+{
+    return validateMessageChecksum(message, messageSize);
+}
+
+#endif
+
